requirements.txt
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
beautifulsoup4
pandas
requests
pandas
numpy
selenium
webdriver-manager
boto3
requests
